refactored sliding text for readability
