+++ styles/Home.module.css
@@ -31,12 +31,6 @@
   text-align: center;
 }
 
-
-.hideOverflow {
-  text-align: center;
-  overflow: hidden;
-}
-
 .shuffleButton {
   padding: 0;
   margin: 0;
